+++ test/sql/basicPragma.sql
@@ -1,0 +1,4 @@
+PRAGMA hat.pants = 'some string';
+PRAGMA pants.pants(+200.00);
+pragma nap.times;
+pragma suit = [no];