+++ test/sql/basicSqlite.sql
@@ -1,0 +1,9 @@
+DETACH hat_db;
+DETACH DATABASE pants_db;
+REINDEX happy_collation;
+REINDEX;
+REINDEX hat_db.pants_table;
+ANALYZE happy_table;
+ANALYZE;
+ANALYZE hat_db.pants_table;
+VACUUM;