+++ test/sql/transactionMisc.sql
@@ -1,0 +1,3 @@
+RELEASE SAVEPOINT happy_place;
+RELEASE sad_place;
+SAVEPOINT bee_time;